azure_function/requirements.txt
azure-iot-hub
azure-cosmos
